refactor: Token 추출 로직 분리
