+++ src/main/java/shopping/auth/AuthInterceptor.java
@@ -3,31 +3,18 @@ package shopping.auth;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
-import shopping.exception.AuthException;
 
 public class AuthInterceptor implements HandlerInterceptor {
 
-    private static final String BEARER = "Bearer ";
-    private static final String AUTHORIZATION = "Authorization";
+    private final TokenExtractor tokenExtractor;
 
-    private final TokenProvider tokenProvider;
-
-    public AuthInterceptor(TokenProvider tokenProvider) {
-        this.tokenProvider = tokenProvider;
+    public AuthInterceptor(TokenExtractor tokenExtractor) {
+        this.tokenExtractor = tokenExtractor;
     }
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        String accessToken = request.getHeader(AUTHORIZATION);
-        validateHeaderExist(accessToken);
-        String email = tokenProvider.getEmail(accessToken.substring(BEARER.length()));
-        request.setAttribute("email", email);
+        tokenExtractor.extractToken(request);
         return HandlerInterceptor.super.preHandle(request, response, handler);
-    }
-
-    private void validateHeaderExist(String accessToken) {
-        if (accessToken == null) {
-            throw new AuthException();
-        }
     }
 }
